+++ src/app/code/community/Shopgate/Cloudapi/Block/Layout.php
@@ -19,7 +19,7 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License, Version 2.0
  */
 
-class Shopgate_Cloudapi_Block_Layout extends Mage_Adminhtml_Block_Template
+class Shopgate_Cloudapi_Block_Layout extends Mage_Core_Block_Template
 {
     /**
      * Print init CSS script into header