+++ src/app/code/community/Shopgate/Cloudapi/Model/Frontend/Observer/OnepageSuccessAction.php
@@ -22,45 +22,59 @@
 class Shopgate_Cloudapi_Model_Frontend_Observer_OnepageSuccessAction
 {
     /**
+     * Prevent observer checkout success
+     */
+    const PREVENT_OBSERVER_CHECKOUT_SUCCESS_KEY = 'prevent_observer_checkout_success';
+
+    /**
+     * Shopgate checkout onepage template path
+     */
+    const CHECKOUT_ONEPAGE_SUCCESS_TEMPLATE_PATH ='shopgate/cloudapi/checkout/success/header.phtml';
+
+    /**
+     * Shopgate Events js path
+     */
+    const SHOPGATE_CONNECT_JS_PATH = 'shopgate/sgConnect.js';
+
+    /**
      * Checks if the order is received from Shopgate API call.
      *
      * @param Varien_Event_Observer $observer
+     * @return $this
      */
     public function execute(Varien_Event_Observer $observer)
     {
-        $session = Mage::getSingleton('checkout/session');
+        $orderIds = $observer->getEvent()->getData('order_ids');
 
-        if ($session->getData(Shopgate_Cloudapi_Helper_Frontend_Checkout::SESSION_IS_SHOPGATE_CHECKOUT)) {
-            $orderIds = $observer->getEvent()->getData('order_ids');
-            if (isset($orderIds[0])) {
-                $newOrderId = $orderIds[0];
-                $layout     = Mage::app()->getLayout();
-                /** @var Shopgate_Cloudapi_Block_PipelineRequest $pipelineRequestBlock */
-                $pipelineRequestBlock = $layout->createBlock('shopgate_cloudapi/pipelineRequest');
-                $pipelineRequestBlock->addMethod(
-                    array(
-                        //@todo-sg: evaluate the right serial
-                        'serial' => Shopgate_Cloudapi_Block_PipelineRequest::PIPELINE_REQUEST_SERIAL,
-                        'name'   => Shopgate_Cloudapi_Block_PipelineRequest::PIPELINE_REQUEST_CREATE_NEW_CART_FOR_CUSTOMER,
-                        'input'  => array(
-                            'orderId' => $newOrderId
-                        )
-                    )
-                );
-                $pipelineRequestBlock->setTemplate('shopgate/cloudapi/pipelineRequest.phtml');
-                /** @var Mage_Page_Block_Html_Head $head */
-                $head = $layout->getBlock('head');
-                $head->addJs('shopgate/pipelineRequest.js');
-                $head->append($pipelineRequestBlock);
-
-                /** @var Shopgate_Cloudapi_Block_Analytics_LogPurchase $analyticsLogPurchaseBlock */
-                $analyticsLogPurchaseBlock = $layout->createBlock('shopgate_cloudapi/analytics_logPurchase');
-                $analyticsLogPurchaseBlock->setOrderId($newOrderId);
-                $analyticsLogPurchaseBlock->setTemplate('shopgate/cloudapi/analyticsLogPurchase.phtml');
-                $head->addJs('shopgate/analyticsLogPurchase.js');
-                $head->append($analyticsLogPurchaseBlock);
-            }
-            $session->unsetData(Shopgate_Cloudapi_Helper_Frontend_Checkout::SESSION_IS_SHOPGATE_CHECKOUT);
+        if ($this->shouldNotExecute($orderIds)) {
+            return $this;
         }
+
+        $newOrderId = $orderIds[0];
+        $layout     = Mage::app()->getLayout();
+
+        $head = $layout->getBlock('head');
+        $head->addJs(self::SHOPGATE_CONNECT_JS_PATH);
+
+        /** @var Shopgate_Cloudapi_Block_Checkout_Onepage_Success $successBlock */
+        $successBlock = $layout->createBlock('shopgate_cloudapi/checkout_onepage_success');
+        $successBlock->setTemplate(self::CHECKOUT_ONEPAGE_SUCCESS_TEMPLATE_PATH);
+        $successBlock->setOrderId($newOrderId);
+
+        $head->append($successBlock);
+
+        Mage::register(self::PREVENT_OBSERVER_CHECKOUT_SUCCESS_KEY, true);
+    }
+
+    /**
+     * @param array $orderIds
+     * @return bool
+     */
+    protected function shouldNotExecute($orderIds)
+    {
+        return !isset($orderIds[0])
+               || !Mage::helper('shopgate_cloudapi/request')->isShopgateRequest() || Mage::registry(
+                self::PREVENT_OBSERVER_CHECKOUT_SUCCESS_KEY
+            );
     }
 }