+++ src/app/code/community/Shopgate/Cloudapi/controllers/V2Controller.php
@@ -50,9 +50,25 @@ class Shopgate_Cloudapi_V2Controller extends Mage_Api_V2_SoapController
     }
 
     /**
+     * All calls related to customers
+     */
+    public function customersAction()
+    {
+        $this->_getServer()->run();
+    }
+
+    /**
      * All calls related to product action
      */
     public function productsAction()
+    {
+        $this->_getServer()->run();
+    }
+
+    /**
+     * All calls related to category information retrieval
+     */
+    public function categoriesAction()
     {
         $this->_getServer()->run();
     }
@@ -67,6 +83,8 @@ class Shopgate_Cloudapi_V2Controller extends Mage_Api_V2_SoapController
 
     /**
      * Initializes the entry point and replaces the path to point to REST
+     *
+     * @throws Mage_Core_Model_Store_Exception
      */
     public function preDispatch()
     {
@@ -93,6 +111,8 @@ class Shopgate_Cloudapi_V2Controller extends Mage_Api_V2_SoapController
      * does not use the REST response Headers & other data.
      * When we return product data, the headers get sent
      * prematurely, so we avoid header manipulation then.
+     *
+     * @throws Zend_Controller_Response_Exception
      */
     public function postDispatch()
     {
@@ -140,7 +160,7 @@ class Shopgate_Cloudapi_V2Controller extends Mage_Api_V2_SoapController
      *
      * @return Shopgate_Cloudapi_Helper_Preflight
      */
-    private final function getCorsHelper()
+    private function getCorsHelper()
     {
         return Mage::helper('shopgate_cloudapi/preflight');
     }