+++ CHANGELOG.md
@@ -8,6 +8,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 - shopgate_order_sources table to flag Shopgate orders
 - UTM parameters for Google Analytics in web checkout
+- Endpoint for retrieving basic customer information
+- Added logic for custom layout handles and removed unnecessary elements from app webview
+- Endpoint stub (blank) for category retrieval; logic to be added later
+### Fixed
+- oAuth refresh token handling 
 
 ## 3.0.0 - 2017-11-23
 ### Added